Changed protection against zero addresses in contracts
