+++ nucypher/blockchain/eth/sol/source/contracts/PolicyManager.sol
@@ -92,9 +92,10 @@ contract PolicyManager is Upgradeable {
     * @param _escrow Escrow contract
     **/
     constructor(MinersEscrow _escrow) public {
-        require(address(_escrow) != address(0));
+        // if the input address is not the MinerEscrow than calling `secondsPerPeriod` will throw error
+        secondsPerPeriod = _escrow.secondsPerPeriod();
+        require(secondsPerPeriod > 0);
         escrow = _escrow;
-        secondsPerPeriod = escrow.secondsPerPeriod();
     }
 
     /**
@@ -501,7 +502,7 @@ contract PolicyManager is Upgradeable {
 
     function verifyState(address _testTarget) public {
         super.verifyState(_testTarget);
-        require(address(uint160(delegateGet(_testTarget, "escrow()"))) == address(escrow));
+        require(address(delegateGet(_testTarget, "escrow()")) == address(escrow));
         require(uint32(delegateGet(_testTarget, "secondsPerPeriod()")) == secondsPerPeriod);
         Policy storage policy = policies[RESERVED_POLICY_ID];
         Policy memory policyToCheck = delegateGetPolicy(_testTarget, RESERVED_POLICY_ID);
