+++ tests/blockchain/eth/contracts/main/mining_adjudicator/test_mining_adjudicator.py
@@ -482,6 +482,10 @@ def test_upgrading(testerchain):
     secret_hash = testerchain.interface.w3.keccak(secret)
     secret2_hash = testerchain.interface.w3.keccak(secret2)
 
+    # Only escrow contract is allowed in MiningAdjudicator constructor
+    with pytest.raises((TransactionFailed, ValueError)):
+        testerchain.interface.deploy_contract('MiningAdjudicator', creator, ALGORITHM_KECCAK256, 1, 2, 3, 4)
+
     # Deploy contracts
     escrow1, _ = testerchain.interface.deploy_contract('MinersEscrowForMiningAdjudicatorMock')
     escrow2, _ = testerchain.interface.deploy_contract('MinersEscrowForMiningAdjudicatorMock')
