+++ tests/blockchain/eth/contracts/main/user_escrow/test_user_escrow.py
@@ -16,13 +16,8 @@ along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 
-import os
-
 import pytest
 from eth_tester.exceptions import TransactionFailed
-from web3.contract import Contract
-
-secret = (123456).to_bytes(32, byteorder='big')
 
 
 @pytest.mark.slow
@@ -358,143 +353,3 @@ def test_policy(testerchain, policy_manager, user_escrow, user_escrow_proxy):
     event_args = events[0]['args']
     assert user == event_args['owner']
     assert 222 == event_args['value']
-
-
-@pytest.mark.slow
-def test_proxy(testerchain, policy_manager, user_escrow):
-    """
-    Test that proxy executes only predefined methods
-    """
-    user = testerchain.interface.w3.eth.accounts[1]
-
-    # Create fake instance of the user escrow contract
-    fake_user_escrow = testerchain.interface.w3.eth.contract(
-        abi=policy_manager.abi,
-        address=user_escrow.address,
-        ContractFactoryClass=Contract)
-
-    # Can't execute method that not in the proxy
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = fake_user_escrow.functions.additionalMethod(1).transact({'from': user})
-        testerchain.wait_for_receipt(tx)
-
-    # And can't send ETH to the user escrow without payable fallback function
-    tx = testerchain.interface.w3.eth.sendTransaction(
-        {'from': testerchain.interface.w3.eth.coinbase, 'to': user, 'value': 1})
-    testerchain.wait_for_receipt(tx)
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = testerchain.interface.w3.eth.sendTransaction(
-            {'from': user, 'to': user_escrow.address, 'value': 1, 'gas_price': 0})
-        testerchain.wait_for_receipt(tx)
-
-
-@pytest.mark.slow
-def test_upgrading(testerchain, token):
-    creator = testerchain.interface.w3.eth.accounts[0]
-    user = testerchain.interface.w3.eth.accounts[1]
-    tx = testerchain.interface.w3.eth.sendTransaction(
-        {'from': testerchain.interface.w3.eth.coinbase, 'to': user, 'value': 1})
-    testerchain.wait_for_receipt(tx)
-
-    secret2 = os.urandom(32)
-    secret_hash = testerchain.interface.w3.keccak(secret)
-    secret2_hash = testerchain.interface.w3.keccak(secret2)
-
-    library_v1, _ = testerchain.interface.deploy_contract('UserEscrowLibraryMockV1')
-    library_v2, _ = testerchain.interface.deploy_contract('UserEscrowLibraryMockV2')
-    linker_contract, _ = testerchain.interface.deploy_contract(
-        'UserEscrowLibraryLinker', library_v1.address, secret_hash)
-    user_escrow_contract, _ = testerchain.interface.deploy_contract(
-        'UserEscrow', linker_contract.address, token.address)
-    # Transfer ownership
-    tx = user_escrow_contract.functions.transferOwnership(user).transact({'from': creator})
-    testerchain.wait_for_receipt(tx)
-    user_escrow_library_v1 = testerchain.interface.w3.eth.contract(
-        abi=library_v1.abi,
-        address=user_escrow_contract.address,
-        ContractFactoryClass=Contract)
-    user_escrow_library_v2 = testerchain.interface.w3.eth.contract(
-        abi=library_v2.abi,
-        address=user_escrow_contract.address,
-        ContractFactoryClass=Contract)
-
-    # Check existed methods and that only user can call them
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = user_escrow_library_v1.functions.firstMethod().transact({'from': creator})
-        testerchain.wait_for_receipt(tx)
-    tx = user_escrow_library_v1.functions.firstMethod().transact({'from': user})
-    testerchain.wait_for_receipt(tx)
-    assert 20 == user_escrow_library_v1.functions.secondMethod().call({'from': user})
-
-    # Nonexistent methods can't be called
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = user_escrow_library_v2.functions.thirdMethod().transact({'from': user})
-        testerchain.wait_for_receipt(tx)
-
-    # Can't send ETH to this version of the library
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = testerchain.interface.w3.eth.sendTransaction(
-            {'from': user, 'to': user_escrow_contract.address, 'value': 1, 'gas_price': 0})
-        testerchain.wait_for_receipt(tx)
-
-    # Only creator can update a library
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = linker_contract.functions.upgrade(library_v2.address, secret, secret2_hash).transact({'from': user})
-        testerchain.wait_for_receipt(tx)
-
-    # Creator must know the secret
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = linker_contract.functions.upgrade(library_v2.address, secret2, secret2_hash).transact({'from': creator})
-        testerchain.wait_for_receipt(tx)
-
-    # Creator can't use the same secret again because it's insecure
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = linker_contract.functions.upgrade(library_v2.address, secret, secret_hash).transact({'from': creator})
-        testerchain.wait_for_receipt(tx)
-
-    assert library_v1.address == linker_contract.functions.target().call()
-    tx = linker_contract.functions.upgrade(library_v2.address, secret, secret2_hash).transact({'from': creator})
-    testerchain.wait_for_receipt(tx)
-    assert library_v2.address == linker_contract.functions.target().call()
-
-    # Method with old signature is not working
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = user_escrow_library_v1.functions.firstMethod().transact({'from': user})
-        testerchain.wait_for_receipt(tx)
-    # Method with old signature that available in new ABI is working
-    assert 15 == user_escrow_library_v1.functions.secondMethod().call({'from': user})
-
-    # New ABI is working
-    assert 15 == user_escrow_library_v2.functions.secondMethod().call({'from': user})
-    tx = user_escrow_library_v2.functions.firstMethod(10).transact({'from': user})
-    testerchain.wait_for_receipt(tx)
-    tx = user_escrow_library_v2.functions.thirdMethod().transact({'from': user})
-    testerchain.wait_for_receipt(tx)
-
-    # And can send and withdraw ETH
-    tx = testerchain.interface.w3.eth.sendTransaction(
-        {'from': user, 'to': user_escrow_contract.address, 'value': 1, 'gas_price': 0})
-    testerchain.wait_for_receipt(tx)
-    assert 1 == testerchain.interface.w3.eth.getBalance(user_escrow_contract.address)
-    # Only user can send ETH
-    with pytest.raises((TransactionFailed, ValueError)):
-        tx = testerchain.interface.w3.eth.sendTransaction(
-            {'from': testerchain.interface.w3.eth.coinbase,
-             'to': user_escrow_contract.address,
-             'value': 1,
-             'gas_price': 0})
-        testerchain.wait_for_receipt(tx)
-    assert 1 == testerchain.interface.w3.eth.getBalance(user_escrow_contract.address)
-
-    rewards = user_escrow_contract.events.ETHWithdrawn.createFilter(fromBlock='latest')
-    user_balance = testerchain.interface.w3.eth.getBalance(user)
-    tx = user_escrow_contract.functions.withdrawETH().transact({'from': user, 'gas_price': 0})
-    testerchain.wait_for_receipt(tx)
-    assert user_balance + 1 == testerchain.interface.w3.eth.getBalance(user)
-    assert 0 == testerchain.interface.w3.eth.getBalance(user_escrow_contract.address)
-
-    events = rewards.get_all_entries()
-    assert 1 == len(events)
-    event_args = events[0]['args']
-    assert user == event_args['owner']
-    assert 1 == event_args['value']
